Rewrite config files
