+++ backend/config/index.js
@@ -1,47 +1,9 @@
-const path = require('path');
-
-const env = process.env.NODE_ENV;
-
-const development = {
-  app: {
-    port: parseInt(process.env.PORT) || 3000,
-    uploads: path.join(__dirname, '../../frontend/public/uploads'),
-  },
-  db: {
-    database: 'dumpach',
-    user: process.env.DB_USER || 'postgres',
-    password: process.env.DB_PASSWORD || '123456',
-    options: {
-      host: process.env.DB_HOST || 'postgres', //Uncomment for docker
-      // host: process.env.DB_HOST || 'localhost',
-      port: parseInt(process.env.DEV_DB_PORT) || 5432,
-      dialect: 'postgres',
-      logging: false,
-    },
-  },
-};
-
-const production = {
-  app: {
-    port: parseInt(process.env.PORT) || 3000,
-    uploads: path.join(__dirname, '../../uploads'),
-  },
-  db: {
-    database: 'dumpach',
-    user: process.env.DB_USER || 'postgres',
-    password: process.env.DB_PASSWORD || '123456',
-    options: {
-      host: process.env.DB_HOST || 'postgres',
-      port: parseInt(process.env.DEV_DB_PORT) || 5432,
-      dialect: 'postgres',
-      logging: false,
-    },
-  },
-};
+const app = require('./app');
+const db = require('./db');
 
 const config = {
-  development,
-  production,
+  app,
+  db,
 };
 
-module.exports = config[env];
+module.exports = config;
